requirements.txt
flask==2.3.2
mysql-connector-python==8.0.33
python-dotenv==1.0.0